requirements.txt
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==0.7
cryptography==41.0.5
Django==4.2.5
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-dotenv==1.4.2
django-filter==23.3
django-js-asset==2.1.0
django-phonenumber-field==7.1.0
django-rest-registration==0.8.2
djangorestframework==3.14.0
drf-yasg==1.21.7
fcm-django==2.0.0
firebase-admin==6.2.0
google-api-core==2.14.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
httplib2==0.22.0
idna==3.4
inflection==0.5.1
msgpack==1.0.7
packaging==23.1
phonenumbers==8.13.21
Pillow==10.0.1
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sorl-thumbnail==12.9.0
sorl-thumbnail-serializer-field==0.2.1
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
